requirements.txt
PyAutoGUI~=0.9.54
openai~=1.6.1
python-dotenv~=1.0.0
matplotlib~=3.8.2
networkx~=3.2.1
scipy~=1.11.4
plotly~=5.18.0